third_party/WebKit/LayoutTests/bluetooth/requestDevice/name-empty-device-from-name-empty-filter.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../resources/mojo-helpers.js"></script>
<script src="../../resources/bluetooth/web-bluetooth-test.js"></script>
<script>
'use strict';
promise_test(() => {
  return setUpPreconnectedDevice({name: ''})
    .then(() => requestDeviceWithKeyDown({filters: [{name: ''}]}))
    .then(device => {
      assert_equals(device.name, '');
    });
}, 'An empty name device can be obtained by empty name filter.');
</script>
